+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.13.2.0] - 2025-XX-XX
+New release of the HyperDbg Debugger.
+
+### Added
+
+### Changed
+- Set variable length (stack frames) for showing the callstack ([link](https://docs.hyperdbg.org/commands/debugging-commands/k))
+
 ## [0.13.1.0] - 2025-04-14
 New release of the HyperDbg Debugger.